Framework for Planet Rotation, Real Time object growth
After meething with Chris, I have the framework for the planet rotation; I can implement that into the game fully,

The game uses real time now to grow the objects, now we will need to make them save and load when players play the game.

Placed objects will now rotate to the center of the planet.