Players can move objects on the planet (v2018.12.6)
Players can move planet objects around. When object is on the planet, tap the "move" icon to start moving the item. Left click and drag to move the item, right click to confirm movement. When players try to move the item, they can't move the camera otherwise it would freak out. This can be improved by disabling this only when players try to move the object, so that they can move the camera to different place and then move the object.

Changed GameManager's "allowCameraMovement" boolean to "stopCameraMovement" as it was misleading. The behaviour remains the same.